refactor: move out bits/bytes enum to utils
