--- src/units.rs
@@ -1,4 +1,0 @@
-pub mod data_units;
-
-// FIXME: Maybe move temperature units here? idk.
-// FIXME: Maybe combine this with utils?
